TIDOC-1023 First iteration sitemap generator for jsduck site.
This is work-in-progress as part of investigating indexing the site using Nutch.
The idea here is to generate an "ugly" sitemap for Nutch to crawl. The search engine would
then return the "ugly" URLs (?_escaped_fragment_=Titanium.API instead of

See ITOSOPS-55 for related discussions.

It may also be possible to dump out plain HTML with metadata and have Nutch or SOLR simply
index from the files. This script also dumps out HTML files for each class, which may be
useful at some point in the future (or not).